requirements.txt

cffi==1.14.0
Click==7.0
dnspython==1.16.0
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-WTF==0.14.3
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
Pillow==2.2.2
pycparser==2.20
pymongo==3.4.0
six==1.14.0
Werkzeug==1.0.0
WTForms==2.2.1
